# buffer: Optimize memory layout for buffer slices so it is better aligned with the 16KB transport socket read size