--- a/test/extensions/transport_sockets/tls/tls_throughput_benchmark.cc
+++ b/test/extensions/transport_sockets/tls/tls_throughput_benchmark.cc
@@ -129,10 +129,15 @@ static void testThroughput(benchmark::State& state) {
 
     state.ResumeTiming();
     uint32_t num_writes = 0;
+    uint32_t num_times_linearize_did_something = 0;
     while (write_buf.length() > 0) {
+      const Buffer::RawSlice initial = write_buf.frontSlice();
       void* mem;
       size_t len = std::min<uint64_t>(write_buf.length(), 16384);
       mem = write_buf.linearize(len);
+      if (initial.mem_ != mem) {
+        ++num_times_linearize_did_something;
+      }
 
       err = SSL_write(client_ssl, mem, len);
       RELEASE_ASSERT(err == static_cast<int>(len),
@@ -142,8 +147,10 @@ static void testThroughput(benchmark::State& state) {
     }
 
     state.counters["writes_per_iteration"] = num_writes;
+    state.counters["num_times_linearize_did_something"] = num_times_linearize_did_something;
   }
   state.counters["throughput"] = benchmark::Counter(bytes_written, benchmark::Counter::kIsRate);
+
   ::close(sockets[0]);
   ::close(sockets[1]);
 }