@antoniovicente 2020-11-20
> My understanding is that we were reading full 16k chunks, but the allocated buffer ended up being 20k because it took 16k plus header/metadata, which rounded up to 20k. But still that was about 4k wasted on every read (20%), so I think it makes a lot of sense to not inline that data.

The first read allocates a buffer with 20kb-64 bytes.  The second 16KB read adds 4032 bytes to the 20kb buffer and reads the remaining 12352 to a 16kb-64 byte buffer.

So buffer is not fragmented, but it results in expensive linearize when writing SSL records later.