@antoniovicente 2020-11-20
> Got it.
> 
> I'm most of the way done writing a simple benchmark around buffer slices and SSL_write() and linearize; it should be easy to adapt it to compare performance.
> 
> I think the performance risk, which may or may not be caught by a micro-benchmark, is the double pointer deref needed to get to the data. One alternative (not sure of the feasibility) is to put the header (pointer and len) directly into the `SliceDeque` instead of pointers to that block.

Yes, moving away from unique pointers in the dequeus would be helpful.  In order to do that we should remove the subclassing used when tracking mutable and immutable slices.  It should just be a relatively straightforward refactoring of the buffer class.