@antoniovicente 2020-11-24
Benchmark based on #14053 

Without this change:
Benchmark                         Time             CPU   Iterations UserCounters...
testThroughput/0/0/10          63.0 us         63.0 us        11162 throughput=2.85891G/s writes_per_iteration=11
testThroughput/1/1/10          63.8 us         63.8 us        10940 throughput=2.82457G/s writes_per_iteration=11
testThroughput/128/1/10        62.9 us         62.9 us        11104 throughput=2.86714G/s writes_per_iteration=11
testThroughput/4096/1/10       63.3 us         63.3 us        11069 throughput=2.84924G/s writes_per_iteration=11
testThroughput/1/2/10          63.9 us         63.9 us        10987 throughput=2.82145G/s writes_per_iteration=11
testThroughput/128/2/10        63.4 us         63.4 us        11091 throughput=2.84219G/s writes_per_iteration=11
testThroughput/4096/2/10       63.3 us         63.3 us        11040 throughput=2.84663G/s writes_per_iteration=11
testThroughput/1/3/10          63.9 us         63.9 us        10936 throughput=2.82096G/s writes_per_iteration=11
testThroughput/128/3/10        63.0 us         63.0 us        11103 throughput=2.86072G/s writes_per_iteration=11
testThroughput/4096/3/10       63.5 us         63.5 us        11022 throughput=2.83686G/s writes_per_iteration=11

With this change:
Benchmark                         Time             CPU   Iterations UserCounters...
testThroughput/0/0/10          58.4 us         58.4 us        11931 throughput=3.08398G/s writes_per_iteration=11
testThroughput/1/1/10          63.5 us         63.5 us        10773 throughput=2.8374G/s writes_per_iteration=11
testThroughput/128/1/10        64.3 us         64.3 us        11112 throughput=2.80334G/s writes_per_iteration=11
testThroughput/4096/1/10       59.3 us         59.3 us        11818 throughput=3.04029G/s writes_per_iteration=11
testThroughput/1/2/10          63.6 us         63.6 us        10817 throughput=2.83287G/s writes_per_iteration=11
testThroughput/128/2/10        64.6 us         64.6 us        11092 throughput=2.78929G/s writes_per_iteration=11
testThroughput/4096/2/10       59.3 us         59.3 us        11827 throughput=3.03833G/s writes_per_iteration=11
testThroughput/1/3/10          63.7 us         63.7 us        10747 throughput=2.8292G/s writes_per_iteration=11
testThroughput/128/3/10        64.4 us         64.4 us        11088 throughput=2.79835G/s writes_per_iteration=11
testThroughput/4096/3/10       59.4 us         59.4 us        11804 throughput=3.03542G/s writes_per_iteration=11