@ggreenway 2020-11-30
> Actually, all pointer accesses are done via Slice::base_ in the base class. There's no change in the number of pointers de-referenced when accessing the contents of slices.

True, but when the data was inlined, the `base_` dereference is almost certainly an L1 cache hit.