@antoniovicente 2020-11-30
> > True, but when the data was inlined, the base_ dereference is almost certainly an L1 cache hit.
> 
> Naively, I would also think that a single allocation vs. 2 (1 for the slice and one for data) would be better also?

It sounds like I should go the extra mile and replace the Slice interface with a struct and get rid of a level of pointers and allocations.  I'll try to look into it sometime this week.